DT-2356 Change the testcontainers example to use a singleton
This allows for the same container to be reused if Spring decides to start a new context - the current model start one for each
