+++ test-app/src/test/kotlin/uk/gov/justice/digital/hmpps/hmppstemplatepackagename/integration/IntegrationTestBase.kt
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.microsoft.applicationinsights.core.dependencies.google.gson.Gson
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.extension.ExtendWith
-import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.boot.test.context.SpringBootTest
@@ -20,13 +19,9 @@ import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
 import org.springframework.test.web.reactive.server.WebTestClient
-import org.testcontainers.containers.localstack.LocalStackContainer
-import org.testcontainers.containers.localstack.LocalStackContainer.Service.SNS
-import org.testcontainers.containers.localstack.LocalStackContainer.Service.SQS
-import org.testcontainers.containers.output.Slf4jLogConsumer
-import org.testcontainers.containers.wait.strategy.Wait
-import org.testcontainers.utility.DockerImageName
 import uk.gov.justice.digital.hmpps.hmppstemplatepackagename.integration.mocks.OAuthExtension
+import uk.gov.justice.digital.hmpps.hmppstemplatepackagename.integration.testcontainers.LocalStackContainer
+import uk.gov.justice.digital.hmpps.hmppstemplatepackagename.integration.testcontainers.LocalStackContainer.setLocalStackProperties
 import uk.gov.justice.digital.hmpps.hmppstemplatepackagename.service.InboundMessageService
 import uk.gov.justice.digital.hmpps.hmppstemplatepackagename.service.OutboundEventsEmitter
 import uk.gov.justice.digital.hmpps.hmppstemplatepackagename.service.OutboundMessageService
@@ -35,8 +30,6 @@ import uk.gov.justice.hmpps.sqs.HmppsQueueService
 import uk.gov.justice.hmpps.sqs.HmppsSqsProperties
 import uk.gov.justice.hmpps.sqs.MissingQueueException
 import uk.gov.justice.hmpps.sqs.MissingTopicException
-import java.io.IOException
-import java.net.ServerSocket
 
 @SpringBootTest(webEnvironment = RANDOM_PORT)
 @Import(IntegrationTestBase.SqsConfig::class, JwtAuthHelper::class)
@@ -154,44 +147,12 @@ abstract class IntegrationTestBase {
   }
 
   companion object {
-    val log = LoggerFactory.getLogger(this::class.java)
+    private val localStackContainer = LocalStackContainer.instance
 
     @JvmStatic
     @DynamicPropertySource
     fun testcontainers(registry: DynamicPropertyRegistry) {
-      startLocalstackIfNotRunning()?.run {
-        getEndpointConfiguration(SNS)
-          .let { it.serviceEndpoint to it.signingRegion }
-          .also {
-            registry.add("hmpps.sqs.localstackUrl") { it.first }
-            registry.add("hmpps.sqs.region") { it.second }
-          }
-      }
+      localStackContainer?.also { setLocalStackProperties(it, registry) }
     }
-
-    private fun startLocalstackIfNotRunning(): LocalStackContainer? {
-      if (localstackIsRunning()) return null
-      val logConsumer = Slf4jLogConsumer(log).withPrefix("localstack")
-      return LocalStackContainer(
-        DockerImageName.parse("localstack/localstack").withTag("0.12.9.1")
-      ).apply {
-        withServices(SNS, SQS)
-        withEnv("HOSTNAME_EXTERNAL", "localhost")
-        withEnv("DEFAULT_REGION", "eu-west-2")
-        waitingFor(
-          Wait.forLogMessage(".*Ready.*", 1)
-        )
-        start()
-        followOutput(logConsumer)
-      }
-    }
-
-    private fun localstackIsRunning(): Boolean =
-      try {
-        val serverSocket = ServerSocket(4566)
-        serverSocket.localPort == 0
-      } catch (e: IOException) {
-        true
-      }
   }
 }
